sync: replace workspace with upstream MuseScore 4.6.5

+++ src/engraving/rw/read400/read400.cpp
@@ -21,9 +21,6 @@
  */
 #include "read400.h"
 
-#include "../editing/mscoreview.h"
-#include "../editing/transpose.h"
-
 #include "dom/audio.h"
 #include "dom/excerpt.h"
 #include "dom/factory.h"
@@ -43,6 +40,7 @@
 #include "dom/harmony.h"
 #include "dom/tie.h"
 #include "dom/breath.h"
+#include "dom/mscoreview.h"
 #include "dom/fret.h"
 #include "dom/dynamic.h"
 #include "dom/hairpin.h"
@@ -51,7 +49,6 @@
 
 #include "engravingerrors.h"
 
-#include "../compat/readstyle.h"
 #include "staffrw.h"
 #include "tread.h"
 
@@ -60,7 +57,7 @@
 using namespace mu::engraving;
 using namespace mu::engraving::read400;
 
-muse::Ret Read400::readScoreFile(Score* score, XmlReader& e, rw::ReadInOutData* data)
+muse::Ret Read400::readScore(Score* score, XmlReader& e, rw::ReadInOutData* data)
 {
     ReadContext ctx(score);
     if (data) {
@@ -69,7 +66,6 @@ muse::Ret Read400::readScoreFile(Score* score, XmlReader& e, rw::ReadInOutData* 
         }
 
         ctx.setPropertiesToSkip(data->propertiesToSkip);
-        ctx.setForcePageMode(data->forcePageMode);
     }
 
     if (!score->isMaster() && data) {
@@ -91,7 +87,7 @@ muse::Ret Read400::readScoreFile(Score* score, XmlReader& e, rw::ReadInOutData* 
         } else if (tag == "Revision") {
             e.skipCurrentElement();
         } else if (tag == "Score") {
-            if (!readScoreTag(score, e, ctx)) {
+            if (!readScore400(score, e, ctx)) {
                 if (e.error() == muse::XmlStreamReader::CustomError) {
                     return make_ret(Err::FileCriticallyCorrupted, e.errorString());
                 }
@@ -119,7 +115,7 @@ muse::Ret Read400::readScoreFile(Score* score, XmlReader& e, rw::ReadInOutData* 
     return muse::make_ok();
 }
 
-bool Read400::readScoreTag(Score* score, XmlReader& e, ReadContext& ctx)
+bool Read400::readScore400(Score* score, XmlReader& e, ReadContext& ctx)
 {
     std::vector<int> sysStaves;
     while (e.readNextStartElement()) {
@@ -161,17 +157,8 @@ bool Read400::readScoreTag(Score* score, XmlReader& e, ReadContext& ctx)
         } else if (tag == "markIrregularMeasures") {
             score->m_markIrregularMeasures = e.readInt();
         } else if (tag == "Style") {
-            // Since version 400, the Style is usually stored in a separate file,
-            // but we also support reading it from the main mscx file.
-            double sp = score->style().value(Sid::spatium).toReal();
-
-            compat::ReadStyleHook::readStyleTag(score, e);
-
-            if (ctx.overrideSpatium()) {
-                ctx.setOriginalSpatium(score->style().spatium());
-                score->style().set(Sid::spatium, sp);
-            }
-            score->m_engravingFont = score->engravingFonts()->fontByName(score->style().styleSt(Sid::musicalSymbolFont).toStdString());
+            // Since version 400, the style is stored in a separate file
+            e.skipCurrentElement();
         } else if (tag == "copyright" || tag == "rights") {
             score->setMetaTag(u"copyright", Text::readXmlText(e, score));
         } else if (tag == "movement-number") {
@@ -224,6 +211,9 @@ bool Read400::readScoreTag(Score* score, XmlReader& e, ReadContext& ctx)
             Spanner* s = toSpanner(Factory::createItemByName(tag, score->dummy()));
             TRead::readItem(s, e, ctx);
             score->addSpanner(s);
+        } else if (tag == "Excerpt") {
+            // Since version 400, the Excerpts are stored in a separate file
+            e.skipCurrentElement();
         } else if (e.name() == "initialPartId") {
             if (score->excerpt()) {
                 score->excerpt()->setInitialPartId(ID(e.readInt()));
@@ -238,38 +228,14 @@ bool Read400::readScoreTag(Score* score, XmlReader& e, ReadContext& ctx)
             }
             e.skipCurrentElement();
         } else if (tag == "Score") {
-            // Since version 400, Excerpts are usually stored in separate files,
-            // but we also support reading them from the main mscx file.
-            ctx.tracks().clear();                 // ???
-            MasterScore* m = score->masterScore();
-            Score* s = m->createScore();
-
-            compat::ReadStyleHook::setupDefaultStyle(s);
-
-            Excerpt* ex = new Excerpt(m);
-            ex->setExcerptScore(s);
-            ctx.setLastMeasure(nullptr);
-
-            Score* curScore = ctx.score();
-            ctx.setScore(s);
-
-            readScoreTag(s, e, ctx);     // recursion
-
-            ctx.setScore(curScore);
-
-            s->linkMeasures(m);
-            ex->setTracksMapping(ctx.tracks());
-            m->addExcerpt(ex);
+            // Since version 400, the Excerpts is stored in a separate file
+            e.skipCurrentElement();
         } else if (tag == "name") {
             String n = e.readText();
             if (!score->isMaster()) {     //ignore the name if it's not a child score
                 score->excerpt()->setName(n, /*saveAndNotify=*/ false);
             }
         } else if (tag == "layoutMode") {
-            if (ctx.forcePageMode()) {
-                e.skipCurrentElement();
-                continue;
-            }
             String s = e.readText();
             if (s == "line") {
                 score->setLayoutMode(LayoutMode::LINE);
@@ -366,10 +332,6 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
         if (tickLen.isZero() || staves == 0) {
             break;
         }
-        if (doScale && !TDuration(tickLen).isValid()) {
-            LOGD("Can't paste: invalid duration %d/%d", tickLen.numerator(), tickLen.denominator());
-            return false;
-        }
 
         Fraction oEndTick = dstTick + oTickLen;
         auto oSpanner = score->spannerMap().findContained(dstTick.ticks(), oEndTick.ticks());
@@ -398,20 +360,6 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                 done = true;
                 break;
             }
-            if (dst->isInsideTupletOnStaff(dstStaffIdx)) {
-                done = true;
-                break;
-            }
-            // Check the time stretch for all measures overlapping the destination range.
-            for (Measure* m = dst->measure(); m && m->tick() < oEndTick; m = m->nextMeasure()) {
-                Fraction mTimeStretch = dst->score()->staff(dstStaffIdx)->timeStretch(m->tick());
-                if (mTimeStretch != Fraction(1, 1)) {
-                    LOGD("Can't paste due to different time stretch ratios (src time stretch: 1/1, dst time stretch: %d/%d)",
-                         mTimeStretch.numerator(), mTimeStretch.denominator());
-                    MScore::setError(MsError::DEST_LOCAL_TIME_SIGNATURE);
-                    return false;
-                }
-            }
 
             while (e.readNextStartElement()) {
                 pasted = true;
@@ -422,15 +370,15 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                 } else if (tag == "transposeDiatonic") {
                     ctx.setTransposeDiatonic(static_cast<int8_t>(e.readInt()));
                 } else if (tag == "voiceOffset") {
-                    Fraction voiceOffset[VOICES];
-                    std::fill(voiceOffset, voiceOffset + VOICES, Fraction(1, 0));
+                    int voiceOffset[VOICES];
+                    std::fill(voiceOffset, voiceOffset + VOICES, -1);
                     while (e.readNextStartElement()) {
                         if (e.name() != "voice") {
                             e.unknown();
                         }
                         voice_idx_t voiceId = static_cast<voice_idx_t>(e.intAttribute("id", -1));
                         assert(voiceId < VOICES);
-                        voiceOffset[voiceId] = Fraction::fromTicks(e.readInt());
+                        voiceOffset[voiceId] = e.readInt();
                     }
                     if (!score->makeGap1(dstTick, dstStaffIdx, tickLen, voiceOffset)) {
                         LOGD() << "cannot make gap in staff " << dstStaffIdx << " at tick " << dstTick.ticks();
@@ -444,17 +392,20 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                 } else if (tag == "Tuplet") {
                     Tuplet* oldTuplet = tuplet;
                     Fraction tick = doScale ? (ctx.tick() - dstTick) * scale + dstTick : ctx.tick();
+                    // no paste into local time signature
+                    if (score->staff(dstStaffIdx)->isLocalTimeSignature(tick)) {
+                        MScore::setError(MsError::DEST_LOCAL_TIME_SIGNATURE);
+                        if (oldTuplet && oldTuplet->elements().empty()) {
+                            delete oldTuplet;
+                        }
+                        return false;
+                    }
                     Measure* measure = score->tick2measure(tick);
                     tuplet = Factory::createTuplet(measure);
                     tuplet->setTrack(ctx.track());
                     TRead::read(tuplet, e, ctx);
                     if (doScale) {
-                        Fraction ticksScaled = tuplet->ticks() * scale;
-                        if (!TDuration(ticksScaled).isValid()) {
-                            LOGD("Can't paste: invalid duration %d/%d", ticksScaled.numerator(), ticksScaled.denominator());
-                            return false;
-                        }
-                        tuplet->setTicks(ticksScaled);
+                        tuplet->setTicks(tuplet->ticks() * scale);
                         tuplet->setBaseLen(tuplet->baseLen().fraction() * scale);
                     }
                     tuplet->setParent(measure);
@@ -495,6 +446,11 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                     TRead::readItem(cr, e, ctx);
                     cr->setSelected(false);
                     Fraction tick = doScale ? (ctx.tick() - dstTick) * scale + dstTick : ctx.tick();
+                    // no paste into local time signature
+                    if (score->staff(dstStaffIdx)->isLocalTimeSignature(tick)) {
+                        MScore::setError(MsError::DEST_LOCAL_TIME_SIGNATURE);
+                        return false;
+                    }
                     if (score->tick2measure(tick)->isMeasureRepeatGroup(dstStaffIdx)) {
                         MeasureRepeat* mr = score->tick2measure(tick)->measureRepeatElement(dstStaffIdx);
                         score->deleteItem(mr);    // resets any measures related to mr
@@ -509,15 +465,11 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                         if (tuplet) {
                             cr->readAddTuplet(tuplet);
                         }
-                        ctx.incTick(cr->actualTicksAt(tick));
+                        ctx.incTick(cr->actualTicks());
                         if (doScale) {
-                            Fraction ticksScaled = cr->ticks() * scale;
-                            if (!TDuration(ticksScaled).isValid()) {
-                                LOGD("Can't paste: invalid duration %d/%d", ticksScaled.numerator(), ticksScaled.denominator());
-                                return false;
-                            }
-                            cr->setTicks(ticksScaled);
-                            cr->setDurationType(cr->durationTypeTicks() * scale);
+                            Fraction d = cr->durationTypeTicks();
+                            cr->setTicks(cr->ticks() * scale);
+                            cr->setDurationType(d * scale);
                             for (Lyrics* l : cr->lyrics()) {
                                 l->setTicks(l->ticks() * scale);
                             }
@@ -530,14 +482,10 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                             TremoloTwoChord* t = chord->tremoloTwoChord();
                             if (t && chord == t->chord2()) {
                                 if (doScale) {
-                                    Fraction ticksScaled = t->durationType().ticks() * scale;
-                                    if (!TDuration(ticksScaled).isValid()) {
-                                        LOGD("Can't paste: invalid duration %d/%d", ticksScaled.numerator(), ticksScaled.denominator());
-                                        return false;
-                                    }
-                                    t->setDurationType(ticksScaled);
+                                    Fraction d = t->durationType().ticks();
+                                    t->setDurationType(d * scale);
                                 }
-                                Fraction tremoloEndTick = tick + chord->actualTicksAt(tick);
+                                Fraction tremoloEndTick = tick + chord->actualTicks();
                                 Fraction measureEndTick = score->tick2measure(tick)->endTick();
                                 if (tremoloEndTick > measureEndTick) {
                                     MScore::setError(MsError::DEST_TREMOLO);
@@ -547,18 +495,13 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                             for (size_t i = 0; i < graceNotes.size(); ++i) {
                                 Chord* gc = graceNotes.at(i);
                                 gc->setGraceIndex(i);
-                                if (gc->vStaffIdx() >= gc->score()->nstaves()) {
-                                    // check if staffMove moves a note to a
-                                    // nonexistent staff
-                                    gc->setStaffMove(0);
-                                }
-                                Transpose::transposeChord(gc, tick);
+                                Score::transposeChord(gc, tick);
                                 chord->add(gc);
                             }
                             graceNotes.clear();
                         }
                         // delete pending ties, they are not selected when copy
-                        if ((tick - dstTick) + cr->actualTicksAt(tick) >= tickLen) {
+                        if ((tick - dstTick) + cr->actualTicks() >= tickLen) {
                             if (cr->isChord()) {
                                 Chord* c = toChord(cr);
                                 for (Note* note: c->notes()) {
@@ -571,12 +514,12 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                             }
                         }
                         // shorten last cr to fit in the space made by makeGap
-                        if ((tick - dstTick) + cr->actualTicksAt(tick) > tickLen) {
+                        if ((tick - dstTick) + cr->actualTicks() > tickLen) {
                             Fraction newLength = tickLen - (tick - dstTick);
                             // check previous CR on same track, if it has tremolo, delete the tremolo
                             // we don't want a tremolo and two different chord durations
                             if (cr->isChord()) {
-                                Segment* s = score->tick2leftSegment(tick - Fraction::eps());
+                                Segment* s = score->tick2leftSegment(tick - Fraction::fromTicks(1));
                                 if (s) {
                                     ChordRest* crt = toChordRest(s->element(cr->track()));
                                     if (!crt) {
@@ -620,7 +563,7 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                     Interval interval = staffDest->transpose(tick);
                     if (!ctx.style().styleB(Sid::concertPitch) && !interval.isZero()) {
                         interval.flip();
-                        Transpose::undoTransposeHarmony(score, harmony, interval);
+                        score->undoTransposeHarmony(harmony, interval);
                     }
 
                     // remove pre-existing chords on this track
@@ -758,18 +701,6 @@ bool Read400::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
 
     for (Score* s : score->scoreList()) {     // for all parts
         s->connectTies();
-
-        for (Spanner* sp : score->unmanagedSpanners()) {
-            if (sp->isLyricsLine() && toLyricsLine(sp)->isDash()) {
-                LyricsLine* line = toLyricsLine(sp);
-                line->setNextLyrics(searchNextLyrics(line->lyrics()->segment(),
-                                                     line->staffIdx(),
-                                                     line->lyrics()->verse(),
-                                                     line->lyrics()->placement()
-                                                     ));
-                line->setTrack2(line->nextLyrics() ? line->nextLyrics()->track() : line->track());
-            }
-        }
     }
 
     if (pasted) {                         //select only if we pasted something
@@ -898,7 +829,7 @@ void Read400::pasteSymbols(XmlReader& e, ChordRest* dst)
                         Interval interval = staffDest->transpose(destTick);
                         if (!ctx.style().styleB(Sid::concertPitch) && !interval.isZero()) {
                             interval.flip();
-                            Transpose::undoTransposeHarmony(score, el, interval);
+                            score->undoTransposeHarmony(el, interval);
                         }
                         el->setParent(harmSegm);
                         score->undoAddElement(el);
