+++ src/engraving/rw/read410/read410.cpp
@@ -21,8 +21,6 @@
  */
 #include "read410.h"
 
-#include "../editing/mscoreview.h"
-#include "../editing/transpose.h"
 #include "../types/types.h"
 
 #include "dom/anchors.h"
@@ -39,6 +37,7 @@
 #include "dom/lyrics.h"
 #include "dom/masterscore.h"
 #include "dom/measurerepeat.h"
+#include "dom/mscoreview.h"
 #include "dom/note.h"
 #include "dom/part.h"
 #include "dom/score.h"
@@ -51,7 +50,6 @@
 
 #include "engravingerrors.h"
 
-#include "../compat/readstyle.h"
 #include "../compat/tremolocompat.h"
 #include "staffread.h"
 #include "tread.h"
@@ -61,7 +59,7 @@
 using namespace mu::engraving;
 using namespace mu::engraving::read410;
 
-muse::Ret Read410::readScoreFile(Score* score, XmlReader& e, rw::ReadInOutData* data)
+muse::Ret Read410::readScore(Score* score, XmlReader& e, rw::ReadInOutData* data)
 {
     ReadContext ctx(score);
     if (data) {
@@ -70,7 +68,6 @@ muse::Ret Read410::readScoreFile(Score* score, XmlReader& e, rw::ReadInOutData* 
         }
 
         ctx.setPropertiesToSkip(data->propertiesToSkip);
-        ctx.setForcePageMode(data->forcePageMode);
     }
 
     if (!score->isMaster() && data) {
@@ -94,7 +91,7 @@ muse::Ret Read410::readScoreFile(Score* score, XmlReader& e, rw::ReadInOutData* 
         } else if (tag == "LastEID") {
             e.skipCurrentElement();
         } else if (tag == "Score") {
-            if (!readScoreTag(score, e, ctx)) {
+            if (!readScore410(score, e, ctx)) {
                 if (e.error() == muse::XmlStreamReader::CustomError) {
                     return make_ret(Err::FileCriticallyCorrupted, e.errorString());
                 }
@@ -122,7 +119,7 @@ muse::Ret Read410::readScoreFile(Score* score, XmlReader& e, rw::ReadInOutData* 
     return muse::make_ok();
 }
 
-bool Read410::readScoreTag(Score* score, XmlReader& e, ReadContext& ctx)
+bool Read410::readScore410(Score* score, XmlReader& e, ReadContext& ctx)
 {
     std::vector<int> sysStaves;
     while (e.readNextStartElement()) {
@@ -171,11 +168,6 @@ bool Read410::readScoreTag(Score* score, XmlReader& e, ReadContext& ctx)
             score->m_showSoundFlags = e.readInt();
         } else if (tag == "markIrregularMeasures") {
             score->m_markIrregularMeasures = e.readInt();
-        } else if (tag == "Style") {
-            // Since version 400, the Style is usually stored in a separate file,
-            // but we also support reading it from the mscx file.
-            compat::ReadStyleHook::readStyleTag(score, e);
-            score->m_engravingFont = score->engravingFonts()->fontByName(score->style().styleSt(Sid::musicalSymbolFont).toStdString());
         } else if (tag == "copyright" || tag == "rights") {
             score->setMetaTag(u"copyright", Text::readXmlText(e, score));
         } else if (tag == "movement-number") {
@@ -243,40 +235,12 @@ bool Read410::readScoreTag(Score* score, XmlReader& e, ReadContext& ctx)
                 ctx.tracks().insert({ strack, dtrack });
             }
             e.skipCurrentElement();
-        } else if (tag == "Score") {
-            // Since version 400, Excerpts are usually stored in separate files,
-            // but we also support reading them from the main mscx file.
-
-            ctx.tracks().clear();             // ???
-            MasterScore* m = score->masterScore();
-            Score* s = m->createScore();
-
-            compat::ReadStyleHook::setupDefaultStyle(s);
-
-            Excerpt* ex = new Excerpt(m);
-            ex->setExcerptScore(s);
-            ctx.setLastMeasure(nullptr);
-
-            Score* curScore = ctx.score();
-            ctx.setScore(s);
-
-            readScoreTag(s, e, ctx);     // recursion
-
-            ctx.setScore(curScore);
-
-            s->linkMeasures(m);
-            ex->setTracksMapping(ctx.tracks());
-            m->addExcerpt(ex);
         } else if (tag == "name") {
             String n = e.readText();
             if (!score->isMaster()) {     //ignore the name if it's not a child score
                 score->excerpt()->setName(n, /*saveAndNotify=*/ false);
             }
         } else if (tag == "layoutMode") {
-            if (ctx.forcePageMode()) {
-                e.skipCurrentElement();
-                continue;
-            }
             String s = e.readText();
             if (s == "line") {
                 score->setLayoutMode(LayoutMode::LINE);
@@ -374,10 +338,6 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
         if (tickLen.isZero() || staves == 0) {
             break;
         }
-        if (doScale && !TDuration(tickLen).isValid()) {
-            LOGD("Can't paste: invalid duration %d/%d", tickLen.numerator(), tickLen.denominator());
-            return false;
-        }
 
         Fraction oEndTick = dstTick + oTickLen;
         auto oSpanner = score->spannerMap().findContained(dstTick.ticks(), oEndTick.ticks());
@@ -410,16 +370,6 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                 done = true;
                 break;
             }
-            // Check the time stretch for all measures overlapping the destination range.
-            for (Measure* m = dst->measure(); m && m->tick() < oEndTick; m = m->nextMeasure()) {
-                Fraction mTimeStretch = dst->score()->staff(dstStaffIdx)->timeStretch(m->tick());
-                if (mTimeStretch != Fraction(1, 1)) {
-                    LOGD("Can't paste due to different time stretch ratios (src time stretch: 1/1, dst time stretch: %d/%d)",
-                         mTimeStretch.numerator(), mTimeStretch.denominator());
-                    MScore::setError(MsError::DEST_LOCAL_TIME_SIGNATURE);
-                    return false;
-                }
-            }
 
             while (e.readNextStartElement()) {
                 pasted = true;
@@ -430,15 +380,15 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                 } else if (tag == "transposeDiatonic") {
                     ctx.setTransposeDiatonic(static_cast<int8_t>(e.readInt()));
                 } else if (tag == "voiceOffset") {
-                    Fraction voiceOffset[VOICES];
-                    std::fill(voiceOffset, voiceOffset + VOICES, Fraction(1, 0));
+                    int voiceOffset[VOICES];
+                    std::fill(voiceOffset, voiceOffset + VOICES, -1);
                     while (e.readNextStartElement()) {
                         if (e.name() != "voice") {
                             e.unknown();
                         }
                         voice_idx_t voiceId = static_cast<voice_idx_t>(e.intAttribute("id", -1));
                         assert(voiceId < VOICES);
-                        voiceOffset[voiceId] = Fraction::fromTicks(e.readInt());
+                        voiceOffset[voiceId] = e.readInt();
                     }
                     if (!score->makeGap1(dstTick, dstStaffIdx, tickLen, voiceOffset)) {
                         LOGD() << "cannot make gap in staff " << dstStaffIdx << " at tick " << dstTick.ticks();
@@ -456,17 +406,20 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                 } else if (tag == "Tuplet") {
                     Tuplet* oldTuplet = tuplet;
                     Fraction tick = doScale ? (ctx.tick() - dstTick) * scale + dstTick : ctx.tick();
+                    // no paste into local time signature
+                    if (score->staff(dstStaffIdx)->isLocalTimeSignature(tick)) {
+                        MScore::setError(MsError::DEST_LOCAL_TIME_SIGNATURE);
+                        if (oldTuplet && oldTuplet->elements().empty()) {
+                            delete oldTuplet;
+                        }
+                        return false;
+                    }
                     Measure* measure = score->tick2measure(tick);
                     tuplet = Factory::createTuplet(measure);
                     tuplet->setTrack(ctx.track());
                     TRead::read(tuplet, e, ctx);
                     if (doScale) {
-                        Fraction ticksScaled = tuplet->ticks() * scale;
-                        if (!TDuration(ticksScaled).isValid()) {
-                            LOGD("Can't paste: invalid duration %d/%d", ticksScaled.numerator(), ticksScaled.denominator());
-                            return false;
-                        }
-                        tuplet->setTicks(ticksScaled);
+                        tuplet->setTicks(tuplet->ticks() * scale);
                         tuplet->setBaseLen(tuplet->baseLen().fraction() * scale);
                     }
                     tuplet->setParent(measure);
@@ -507,6 +460,11 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                     TRead::readItem(cr, e, ctx);
                     cr->setSelected(false);
                     Fraction tick = doScale ? (ctx.tick() - dstTick) * scale + dstTick : ctx.tick();
+                    // no paste into local time signature
+                    if (score->staff(dstStaffIdx)->isLocalTimeSignature(tick)) {
+                        MScore::setError(MsError::DEST_LOCAL_TIME_SIGNATURE);
+                        return false;
+                    }
                     if (score->tick2measure(tick)->isMeasureRepeatGroup(dstStaffIdx)) {
                         MeasureRepeat* mr = score->tick2measure(tick)->measureRepeatElement(dstStaffIdx);
                         score->deleteItem(mr);    // resets any measures related to mr
@@ -523,13 +481,9 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                         }
                         ctx.incTick(cr->actualTicksAt(tick));
                         if (doScale) {
-                            Fraction ticksScaled = cr->ticks() * scale;
-                            if (!TDuration(ticksScaled).isValid()) {
-                                LOGD("Can't paste: invalid duration %d/%d", ticksScaled.numerator(), ticksScaled.denominator());
-                                return false;
-                            }
-                            cr->setTicks(ticksScaled);
-                            cr->setDurationType(cr->durationTypeTicks() * scale);
+                            Fraction d = cr->durationTypeTicks();
+                            cr->setTicks(cr->ticks() * scale);
+                            cr->setDurationType(d * scale);
                             for (Lyrics* l : cr->lyrics()) {
                                 l->setTicks(l->ticks() * scale);
                             }
@@ -553,14 +507,10 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                             TremoloTwoChord* tremolo = chord->tremoloTwoChord();
                             if (tremolo && chord == tremolo->chord2()) {
                                 if (doScale) {
-                                    Fraction ticksScaled = tremolo->durationType().ticks() * scale;
-                                    if (!TDuration(ticksScaled).isValid()) {
-                                        LOGD("Can't paste: invalid duration %d/%d", ticksScaled.numerator(), ticksScaled.denominator());
-                                        return false;
-                                    }
-                                    tremolo->setDurationType(ticksScaled);
+                                    Fraction d = tremolo->durationType().ticks();
+                                    tremolo->setDurationType(d * scale);
                                 }
-                                Fraction tremoloEndTick = tick + chord->actualTicksAt(tick);
+                                Fraction tremoloEndTick = tick + chord->actualTicks();
                                 Fraction measureEndTick = score->tick2measure(tick)->endTick();
                                 if (tremoloEndTick > measureEndTick) {
                                     MScore::setError(MsError::DEST_TREMOLO);
@@ -570,18 +520,13 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                             for (size_t i = 0; i < graceNotes.size(); ++i) {
                                 Chord* gc = graceNotes.at(i);
                                 gc->setGraceIndex(i);
-                                if (gc->vStaffIdx() >= gc->score()->nstaves()) {
-                                    // check if staffMove moves a note to a
-                                    // nonexistent staff
-                                    gc->setStaffMove(0);
-                                }
-                                Transpose::transposeChord(gc, tick);
+                                Score::transposeChord(gc, tick);
                                 chord->add(gc);
                             }
                             graceNotes.clear();
                         }
                         // delete pending ties, they are not selected when copy
-                        if ((tick - dstTick) + cr->actualTicksAt(tick) >= tickLen) {
+                        if ((tick - dstTick) + cr->actualTicks() >= tickLen) {
                             if (cr->isChord()) {
                                 Chord* c = toChord(cr);
                                 for (Note* note: c->notes()) {
@@ -594,12 +539,12 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                             }
                         }
                         // shorten last cr to fit in the space made by makeGap
-                        if ((tick - dstTick) + cr->actualTicksAt(tick) > tickLen) {
+                        if ((tick - dstTick) + cr->actualTicks() > tickLen) {
                             Fraction newLength = tickLen - (tick - dstTick);
                             // check previous CR on same track, if it has tremolo, delete the tremolo
                             // we don't want a tremolo and two different chord durations
                             if (cr->isChord()) {
-                                Segment* s = score->tick2leftSegment(tick - Fraction::eps());
+                                Segment* s = score->tick2leftSegment(tick - Fraction::fromTicks(1));
                                 if (s) {
                                     ChordRest* crt = toChordRest(s->element(cr->track()));
                                     if (!crt) {
@@ -643,7 +588,7 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
                     Interval interval = staffDest->transpose(tick);
                     if (!ctx.style().styleB(Sid::concertPitch) && !interval.isZero()) {
                         interval.flip();
-                        Transpose::undoTransposeHarmony(score, harmony, interval);
+                        score->undoTransposeHarmony(harmony, interval);
                     }
 
                     // remove pre-existing chords on this track
@@ -783,18 +728,6 @@ bool Read410::pasteStaff(XmlReader& e, Segment* dst, staff_idx_t dstStaff, Fract
 
     for (Score* s : score->scoreList()) {     // for all parts
         s->connectTies();
-
-        for (Spanner* sp : score->unmanagedSpanners()) {
-            if (sp->isLyricsLine() && toLyricsLine(sp)->isDash()) {
-                LyricsLine* line = toLyricsLine(sp);
-                line->setNextLyrics(searchNextLyrics(line->lyrics()->segment(),
-                                                     line->staffIdx(),
-                                                     line->lyrics()->verse(),
-                                                     line->lyrics()->placement()
-                                                     ));
-                line->setTrack2(line->nextLyrics() ? line->nextLyrics()->track() : line->track());
-            }
-        }
     }
 
     if (pasted) {                         //select only if we pasted something
@@ -1038,7 +971,7 @@ void Read410::pasteSymbols(XmlReader& e, ChordRest* dst)
                     Interval interval = staffDest->transpose(destTick);
                     if (!ctx.style().styleB(Sid::concertPitch) && !interval.isZero()) {
                         interval.flip();
-                        Transpose::undoTransposeHarmony(score, el, interval);
+                        score->undoTransposeHarmony(el, interval);
                     }
                     el->setParent(seg);
                     score->undoAddElement(el);
